fix: 서버에서 만드는 seed에 0이 들어갈 수 있는 부분 수정 및 seed 타입이 uint64로 잘못되어있는 부분수정

+++ server/handler/leaderboard.go
@@ -88,7 +88,7 @@ func PostLeaderboardV1(c echo.Context) BaseResponse {
 		}
 	}
 	seed := sess.Values["seed"]
-	seed = seed.(uint64)
+	seed = seed.(int32)
 	if leaderboard.Seed != seed {
 		return BaseResponse{
 			http.StatusBadRequest,
